重新设计资源管理器TResManager。更改GUI 坐标类型值从integer到Single。 对地图进行代码添加

+++ trunk/Client/DrawEx.pas
@@ -4,10 +4,10 @@ interface
 uses
 MondoZenGL,Texture;
 //X Y ÎªÆÁÄ»µÄÎ»ÖÃ
-Procedure DrawTexture2Canvas(Canvas:TMZcanvas;Texture:TMZTexture;X,Y:integer);
+Procedure DrawTexture2Canvas(Canvas:TMZcanvas;Texture:TMZTexture;X,Y:Single);
 implementation
 
-Procedure DrawTexture2Canvas(Canvas:TMZcanvas;Texture:TMZTexture;X,Y:integer);
+Procedure DrawTexture2Canvas(Canvas:TMZcanvas;Texture:TMZTexture;X,Y:Single);
 var
 coor:TMZTextureCoordinates;
 rect:TMZRect;
